requirements.txt
alabaster==0.7.9
Babel==2.3.4
colorama==0.3.7
decorator==4.0.10
Django==1.10.2
django-ajax-selects==1.5.1
django-extensions==1.7.4
docutils==0.12
imagesize==0.7.1
ipython==5.1.0
ipython-genutils==0.1.0
Jinja2==2.8
MarkupSafe==0.23
pickleshare==0.7.4
prompt-toolkit==1.0.8
Pygments==2.1.3
pytz==2016.7
simplegeneric==0.8.1
six==1.10.0
snowballstemmer==1.2.1
Sphinx==1.4.8
traitlets==4.3.1
wcwidth==0.1.7
win-unicode-console==0.5
